refactor access level 2

+++ Aero.Application/Interfaces/IAccessLevelService.cs
@@ -6,8 +6,8 @@ namespace Aero.Application.Interface
     public interface IAccessLevelService
     {
         Task<ResponseDto<IEnumerable<AccessLevelDto>>> GetAsync();
-        Task<ResponseDto<IEnumerable<AccessLevelDto>>> GetByLocationIdAsync(short location);
-        Task<ResponseDto<Pagination<AccessLevelDto>>> GetPaginationAsync(PaginationParamsWithFilter param,short location);
+        Task<ResponseDto<IEnumerable<AccessLevelDto>>> GetByLocationIdAsync(int location);
+        Task<ResponseDto<Pagination<AccessLevelDto>>> GetPaginationAsync(PaginationParamsWithFilter param, int location);
         Task<ResponseDto<AccessLevelDto>> GetByIdAsync(int id);
         Task<ResponseDto<AccessLevelDto>> CreateAsync(CreateAccessLevelDto dto);
         Task<ResponseDto<AccessLevelDto>> DeleteAsync(int id);
